% Generated by roxygen2 (4.0.1): do not edit by hand
\name{circle_points}
\alias{circle_points}
\title{Surround a set of points with a circle of points at a pre-specified
radius.}
\usage{
circle_points(x, y, np = 6, r = 0.5)
}
\arguments{
\item{x}{vector of x coordinates for points to surround}

\item{y}{vector of y coordinates for points to surround}

\item{np}{number of external points}

\item{r}{radius of circle, measured in distance from center of the
set of points.}
}
\description{
Surround a set of points with a circle of points at a pre-specified
radius.
}

