% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cluster_points.hsmap}
\alias{cluster_points.hsmap}
\title{cluster_points.hsmap}
\usage{
\method{cluster_points}{hsmap}(x, cutoff = 0.95)
}
\arguments{
\item{x}{fitted hsmap object}

\item{cutoff}{pvalue cutoff for cluster membership}
}
\description{
Get points within cluster areas, defined by threshold
}

