% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dbm_score_rr}
\alias{dbm_score_rr}
\title{Generate a scored map of the local proportion of cases.}
\usage{
dbm_score_rr(df, dim = 100, h = 0.005, p = 0.005, bounds = NULL)
}
\arguments{
\item{df}{data frame containing (x,y) coordinates of cases and case control status (z: 0,1)}
}
\description{
Returns a data frame with x and y coordinates of grid squares of
pre-specified dimension and the dbm score for each square.
}

