% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fixed_point_cdf}
\alias{fixed_point_cdf}
\title{Calculate an empirical cdf of point distances relative to a fixed point.}
\usage{
fixed_point_cdf(x, y, cx, cy)
}
\arguments{
\item{x}{vector of x coordinates of data points}

\item{y}{vector of y coordinates of data points}

\item{cx}{fixed point x coordinate}

\item{cy}{fixed point y coordinate}
}
\description{
Calculate an empirical cdf of point distances relative to a fixed point.
}

