% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-hotspot-model.R
\name{generate_hotspot_model}
\alias{generate_hotspot_model}
\title{generate_hotspot_model}
\usage{
generate_hotspot_model(n, alpha = 0.1, sd0 = 0.1, ac_type = "moran",
  niters = 1, plot = FALSE)
}
\arguments{
\item{n}{Number of observations to generate}

\item{alpha}{Strength of spatial autocorrelation}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}

\item{ac_type}{Type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-org}=\code{go})}

\item{niters}{Number of successive layers of spatial autocorrelation}

\item{plot}{If TRUE, produces a plot of rank--scale distributions}
}
\value{
A matrix of two columns containing sorted and scaled versions of
\enumerate{
  \item z = raw values
  \item ac = associated spatial autocorrelation statistics
}
}
\description{
Uses the parameters returned by \code{fit_hotspot_model} to generate
rank-scale distributions both of raw values and associated spatial
autocorrelation statistics.
}

