% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-neighbours.R
\name{get_neighbours}
\alias{get_neighbours}
\title{get_neighbours}
\usage{
get_neighbours(x, y)
}
\arguments{
\item{x}{Vector of x-coordinates}

\item{y}{Vector of y-coordinates}

\item{plot}{Produce a plot of the Delaunay triangulation}
}
\value{
A list of Delaunay neighbours of each point in (x, y)
}
\description{
Determines neighbours of points (x, y) using Delaunay triangulation
}
\examples{
nbs <- get_neighbours (runif (10), runif (10))

}

