% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grid_from_points}
\alias{grid_from_points}
\title{Create a square grid of points of the pre-specified dimension}
\usage{
grid_from_points(x, y, dim)
}
\arguments{
\item{x}{vector of x coordinates of data points}

\item{y}{vector of y coordinates of data points}

\item{dim}{dimension of resulting square grid}
}
\description{
Create a square grid of points of the pre-specified dimension
}

