% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hotspot_area}
\alias{hotspot_area}
\title{Given a map with a single hotspot, extract the squares with an approximate p-value
greater than or equal to a given threshhold}
\usage{
hotspot_area(df, datad, pv)
}
\arguments{
\item{df}{color dataframe output from make_hotspot}

\item{datad}{data frame containing data points}

\item{pv}{threshhold p-value}
}
\description{
Given a map with a single hotspot, extract the squares with an approximate p-value
greater than or equal to a given threshhold
}

