% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hotspot_map}
\alias{hotspot_map}
\title{Generate a hotspot map}
\usage{
hotspot_map(in_df, fn, color_samples = 100, p = 0.1, dim = in_dim,
  bounds = NULL, pbar = TRUE)
}
\arguments{
\item{in_df}{data frame consisting of x and y coordinates of input data, and any
other information used by the mapping function.}

\item{fn}{mapping function that takes a data frame as input and returns a data frame #' with a grid of points of size dim x dim.}

\item{bounds}{gives the coordinates of a bounding box to restrict the map to (otherwise uses the extreme values of points in in_df)}

\item{pbar}{show a progress bar while calculating the color scale.}
}
\description{
This function takes a function for computing some kind of spatial risk,
calculates spatial densities and generates a color scale indicating which areas
have a statistically significant concentration of cases.
}

