% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ives.R
\name{ives}
\alias{ives}
\title{ives}
\usage{
ives(size = 10, nt = 1000, sd0 = 0.1, alpha = c(0.1, 0.1),
  spatial = FALSE, rook = TRUE, seed)
}
\arguments{
\item{size}{Size of the square grid on which to generate model. Total number
of points is size ^ 2}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate final modelled values}

\item{sd0}{Standard deviation of distributions from which values of 'r' and
's' are drawn.}

\item{alpha}{Vector of two components respectively specifying the strength of
autocorrelation in time and space.}

\item{spatial}{If TRUE, spatial autocorrelation is calculated by assuming
movement along locally maximal gradients, otherwise it is equivalent to
neutral model}

\item{rook}{If FALSE, use queen contiguities}

\item{seed}{Random seed}
}
\value{
A list of:
\itemize{
\item \code{z} = Vector of simulated data values
\item \code{nbs} = An \code{spdep::nb} list of neighbours of all points
}
}
\description{
Applies the model of Ives & Klopfer (Ecology 1997) on a square grid of (size,
size)
}
\examples{
dat <- ives (size=10)

}
\seealso{
\code{neutral_hotspots}
}

