% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ives2D.R
\name{ives2d}
\alias{ives2d}
\title{ives2d}
\usage{
ives2d(size = 10, nt = 1000, sd0 = 0.1, alpha = c(0.1, 0.1), seed)
}
\arguments{
\item{size}{Size of the square grid on which to generate model. Total number
of points is size ^ 2}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate final modelled values}

\item{sd0}{Standard deviation of distributions from which values of 'r' and
's' are drawn.}

\item{alpha}{Vector of two components respectively specifying the strength of
autocorrelation in time and space.}

\item{seed}{Random seed}
}
\value{
A matrix of (size, size)
}
\description{
Generates a 2D surface using the temporal autocorrelation model of Ives &
Klopfer (Ecology 1997).
}

