% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morani.R
\name{morani}
\alias{morani}
\title{morani}
\usage{
morani(z0)
}
\arguments{
\item{z0}{Square data matrix}
}
\value{
A vector of values of Moran's I statistic for each point in the grid
}
\description{
Calculates Moran's I statistic for spatial association between 5 rectangular
neighbouring points on a regular grid
}
\seealso{
\code{gearyc}, \code{getisord}
}

