% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral-hotspots.R
\name{neutral_hotspots}
\alias{neutral_hotspots}
\title{neutral_hotspots}
\usage{
neutral_hotspots(nbs, wts, alpha = c(0.1, 0.1), nt = 100, sd0 = 0.1, seed)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs. Uniform weights used if not given.}

\item{alpha}{Vector of two components respectively specifying the strength of
autocorrelation in time and space.}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate final modelled values}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}

\item{seed}{Random seed}
}
\value{
A vector of hotspot values sorted from high to low
}
\description{
Implements neutral model in two dimensions
}
\examples{
nbs <- ives (size=10)$nbs
z <- neutral_hotspots (nbs=nbs)

}
\seealso{
\code{ives}
}

