% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral-hotspots.R
\name{order_one}
\alias{order_one}
\title{order_one}
\usage{
order_one(sigma, n, ntrials = 1000)
}
\arguments{
\item{sigma}{Standard deviation of normal distribution}

\item{n}{Number of samples of normal distribution}

\item{ntrials}{Number of trials over which to average order statistics}
}
\value{
First order statistic
}
\description{
First order statistic for normal distribution.
}
