% Generated by roxygen2 (4.0.1): do not edit by hand
\name{random_hotspot}
\alias{random_hotspot}
\title{Generate a random hotspot.}
\usage{
random_hotspot(x, y, h2, in_p, out_p)
}
\arguments{
\item{x}{vector of x coordinates for input points}

\item{y}{vector of y coordinates for input points}

\item{h2}{width of hotspot window}

\item{in_p}{risk of being in case group inside hotspot}

\item{out_p}{risk of being in case group outside hotspot}
}
\description{
Given a set of points, generate a square hotspot with width h in the middle.
Useful for testing whether a map based on a given set of points will work.
Specify the within-spot risk (in_p) and outside-spot risk (out_p), so that the
relative risk (RR) is equal to in_p / out_p.
}

