% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ac_stats}
\alias{rcpp_ac_stats}
\title{rcpp_ac_stats}
\usage{
rcpp_ac_stats(z, nbs, wts, ac_type)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{ac_type}{Character string specifying type of aucorrelation
(\code{moran}, \code{geary}, or \code{getis-ord}).}

\item{x}{Corresponding vector of values}
}
\value{
A vector of sorted spatial autocorrelation statistics scaled between
zero and one.
}
\description{
Computes spatial autocorrelation statistics for a given input matrix
}
