% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_neutral_hotspots}
\alias{rcpp_neutral_hotspots}
\title{rcpp_neutral_hotspots}
\usage{
rcpp_neutral_hotspots(nbs, wts, nbsi, alpha, sd0, log_scale, niters, ac_type)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point.}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs.}

\item{nbsi}{List of matrices as returned from \code{get_nbsi}. each element
of which contains the i-th nearest neighbour to each point.}

\item{alpha}{Strength of spatial autocorrelation}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}

\item{log_scale}{If TRUE, raw hotspot values are log-transformed}

\item{niters}{Number of iterations of spatial autocorrelation}

\item{ac_type}{Character string specifying type of aucorrelation
(\code{moran}, \code{geary}, or \code{getis-ord}).}
}
\value{
A vector of simulated values of same size as \code{nbs}.
}
\description{
Implements neutral model in two dimensions
}
