% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_neutral_hotspots_ntests}
\alias{rcpp_neutral_hotspots_ntests}
\title{rcpp_neutral_hotspots_ntests}
\usage{
rcpp_neutral_hotspots_ntests(
  nbs,
  wts,
  nbsi,
  alpha,
  sd0,
  niters,
  ac_type,
  log_scale,
  ntests
)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point.}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs.}

\item{nbsi}{List of matrices as returned from \code{get_nbsi}. each element
of which contains the i-th nearest neighbour to each point.}

\item{alpha}{Strength of spatial autocorrelation}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}

\item{ac_type}{Character string specifying type of aucorrelation
(\code{moran}, \code{geary}, or \code{getis-ord}).}

\item{ntests}{Number of tests used to obtain average values}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate final modelled values}
}
\value{
A matrix of dimension (size, 2), with first column containing
sorted and re-scaled hotspot values, and second column containing sorted and
re-scaled spatial autocorrelation statistics.
}
\description{
Performs repeated neutral tests to yield average distributions of both
hotspot values and spatial autocorrelation statistics.
}
