% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-tests.R
\name{run_tests}
\alias{run_tests}
\title{run_tests}
\usage{
run_tests(nbs, alpha = c(0.1, 0.1), nt = 100, ntests = 100,
  neutral = FALSE, ac_type = "moran", z, seed)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate test field}

\item{ntests}{Number of tests to run, with statistics calculated from the
mean of \code{ntests}}

\item{neutral}{If TRUE, uses a neutral field for the tests rather than a
structured field generated by \code{ives}}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{z}{Vector of observed values to be tested. If not given, will
be generated with specified values of \code{alpha} and \code{size=10}}

\item{seed}{Random seed}
}
\value{
Nothing (dumps statistics to screen)
}
\description{
Tests observed data (\code{z}) against a series of neutral models
}
\examples{
\dontrun{
run_tests () # should give low p-values
run_tests (neutral=TRUE) # should give high p-values
}

}
\seealso{
\code{test_hotspots}
}

