% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1d.R
\name{test1da}
\alias{test1da}
\title{test1d}
\usage{
test1da(ydat, alpha = c(0.1, 0.1))
}
\arguments{
\item{ydat}{Square matrix of observed values to be tested}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}
}
\value{
A vector of four values as estimated by the neutral model:
\begin{itemize}
  \item Temporal autocorrelation coefficient
  \item Spatial autocorrelation coefficient
  \item Number of successive layers of spatio-temporal autocorrelation
  required to reproduce statistical properties of observed data
  \item Absolute difference between observed and modelled rank--scale
  distributions
\end{itemize}
}
\description{
Tests an observed matrix of data (\code{ydat}) against a one-dimensional
neutral model
}
\seealso{
\code{test2d}
}

