% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2d.R
\name{test2d}
\alias{test2d}
\title{test2d}
\usage{
test2d(ydat, size = 10, alpha = c(0.1, 0.1), sann = FALSE,
  separate = FALSE, plot = FALSE)
}
\arguments{
\item{ydat}{Square matrix of observed values to be tested}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}
}
\value{
A vector of four values as estimated by the neutral model:
\begin{itemize}
  \item Temporal autocorrelation coefficient
  \item Spatial autocorrelation coefficient
  \item Number of successive layers of spatio-temporal autocorrelation
  required to reproduce statistical properties of observed data
  \item Absolute difference between observed and modelled rank--scale
  distributions
\end{itemize}
}
\description{
Tests an observed matrix of data (\code{ydat}) against a two-dimensional
neutral model
}
\seealso{
\code{test1d}
}

