% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R
\name{testn}
\alias{testn}
\title{testn}
\usage{
testn(nbs, wts, alpha = 0.1, sd0 = 0.1, niters = 1, ac_type = "moran",
  log_scale = TRUE, ntests = 100, seed)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs. Uniform weights used if not given.}

\item{alpha}{strength of spatial autocorrelation}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}

\item{niters}{Number of successive layers of spatial autocorrelation}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{log_scale}{If TRUE, raw hotspot values are log-transformed}

\item{ntests}{Number of tests over which to generate an average result}

\item{seed}{Random seed}
}
\value{
A vector of hotspot values sorted from high to low
}
\description{
Implements neutral model of hotspot values and associated autocorrelation
statustics. Current version is a simply internal R loop, while
\code{neutral_hotspots_ntests} is a parallel version of exactly the same
thing.
}
\section{Note}{
 Just a wrapper for test calls to \code{rcpp_neutral_hotspots}
and \code{rcpp_neutral_hotspots_ntests}
}

