% Generated by roxygen2 (4.0.1): do not edit by hand
\name{within_cluster_rr}
\alias{within_cluster_rr}
\title{within_cluster_rr}
\usage{
within_cluster_rr(x, cutoff = 0.95)
}
\description{
returns a fitted binomial glm predicting the likelihood of being a case, conditional on cluster membership
}

