% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-hotspot-model.R
\name{fit_hotspot_model}
\alias{fit_hotspot_model}
\title{fit_hotspot_model}
\usage{
fit_hotspot_model(z, nbs, wts, ac_type = "moran", ntests = 100,
  verbose = FALSE, plot = FALSE)
}
\arguments{
\item{z}{Vector of observed values to be tested}

\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs. Uniform weights used if not given.}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{ntests}{Number of repeats of neutral model used to calculate mean
rank--scale distribution}

\item{verbose}{If TRUE, dump progress details to screen}

\item{plot}{If TRUE, produces a plot of rank--scale distributions}
}
\value{
A vector of three values as estimated by the neutral model:
\enumerate{
  \item sd0 = standard deviation of normal distribution
  \item alpha = temporal autocorrelation coefficient
  \item niters = number of iterations of spatial autocorrelation
}
}
\description{
Fits a neutral model to an observed series of hotspot values in terms of
standard deviation of environmental variables, spatial autocorrelation
parameters, and number of iterations of spatial autocorrelation.
}
\section{Note}{

Fitting these neutral models is **not** a standard optimisation problem
because the models are very noisy. Although \code{optim} with
\code{method="SANN"} may be used, it often generates extremely large values
for \code{alpha} (for example, > 10). \code{DEoptim} could also be applied,
yet in generally does not explore anything useful---if given starting
parameters, it will generally remain exactly in that place.

The approach employed here reflects the comment of
https://stat.ethz.ch/pipermail/r-help/2015-May/428751.html
through simply producing regular series, fitting loess models, and taking the
corresponding minima.
}
\examples{
\dontrun{
xy <- cbind (rep (seq (size), each=size), rep (seq (size), size))
dhi <- 1 # for rook; dhi=1.5 for queen
nbs <- spdep::dnearneigh (xy, 0, dhi)
z <- runif (length (nbs))
test <- fit_hotspot_model (z=z, nbs=dat$nbs, alpha=0.1, sd=0.1, ntests=100)
}

}

