% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearyc.R
\name{gearyc}
\alias{gearyc}
\title{gearyc}
\usage{
gearyc(z0)
}
\arguments{
\item{z0}{Square data matrix}
}
\value{
A sorted vector of values of Geary's C statistic re-scaled between 0
and 1
}
\description{
Calculates Geary's C statistic for spatial association between 5 rectangular
neighbouring points on a regular grid
}
\seealso{
\code{morani}, \code{getisord}
}

