% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getisord.R
\name{getis_ord}
\alias{getis_ord}
\title{getisord}
\usage{
getis_ord(z0)
}
\arguments{
\item{z0}{Square data matrix}
}
\value{
A sorted vector of values of the Getis-Ord statistic re-scaled
between 0 and 1
}
\description{
Calculates the Getis-Ord statistic for spatial association between 5
rectangular neighbouring points on a regular grid
}
\seealso{
\code{morani}, \code{morani}
}

