% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral2d.R
\name{neutral2d}
\alias{neutral2d}
\title{neutral2d}
\usage{
neutral2d(size = 10, alpha = c(0.1, 0.1), n = 100, sd0 = 0.1)
}
\arguments{
\item{size}{Size of the square grid on which to generate model. Total number
of points is size ^ 2}

\item{alpha}{Vector of two components respectively specifying the strength of
autocorrelation in time and space.}

\item{n}{Number of successive layers of temporal and spatial autocorrelation
used to generate final modelled values}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}
}
\value{
A vector of hotspot values sorted from high to low
}
\description{
Implements neutral model in two dimensions
}
\examples{
y <- neutral2d ()

}
\seealso{
\code{neutral1d}
}

