% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral-hotspots-ntests.R
\name{neutral_hotspots_ntests}
\alias{neutral_hotspots_ntests}
\title{neutral_hotspots_ntests}
\usage{
neutral_hotspots_ntests(nbs, wts, alpha = 0.1, sd0 = 0.1, ntests = 1000,
  ac_type = "moran", seed)
}
\arguments{
\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs. Uniform weights used if not given.}

\item{alpha}{Strength of spatial autocorrelation}

\item{sd0}{Standard deviation of truncated normal distribution used to model
environmental variation (with mean of 1)}

\item{ntests}{Number of repeats of neutral model used to calculate mean
rank--scale distribution}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{seed}{Random seed}
}
\value{
A vector of hotspot values sorted from high to low
}
\description{
Averages results of a number of neutral models
}
\examples{
nbs <- ives (size=10)$nbs
z <- neutral_hotspots (nbs=nbs)

}
\seealso{
\code{ives}
}

