% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-values.R
\name{p_values}
\alias{p_values}
\title{p_values}
\usage{
p_values(z, nbs, wts, alpha = c(0.1, 0.1), nt = 100, ntests = 1000,
  ac_type = "moran")
}
\arguments{
\item{z}{Vector of observed values to be tested}

\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{wts}{Weighting factors for each neighbour; must have same length as
nbs. Uniform weights used if not given.}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate test field}

\item{ntests}{Number of tests to run, with statistics calculated from the
mean of \code{ntests}}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}
}
\value{
Nothing (dumps statistics to screen)
}
\description{
Tests observed data (\code{z}) against a series of neutral models
}

