% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.hsmap}
\alias{plot.hsmap}
\title{plot.hsmap}
\usage{
\method{plot}{hsmap}(x, point = TRUE, pointtype = c(20, 3),
  xlab = "Longitude", ylab = "Latitude", map = NULL, fname = NULL)
}
\arguments{
\item{x}{object of type hsmap}
}
\description{
plot.hsmap
}

