% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{run_tests}
\usage{
run_tests(size = 10, alpha = c(0.1, 0.1), ntests = 100, ydat, seed)
}
\arguments{
\item{size}{Size of the square grid on which to generate model. Total number
of points is size ^ 2. (Has no effect if \code{ydat} is given, in which case
value is determined by its dimensions)}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{ntests}{Number of tests to run, with statistics calculated from the
mean of \code{ntests}}

\item{ydat}{Square matrix of observed values to be tested. If not given, will
be generated with specified values of \code{size} and \code{alpha}}

\item{seed}{Random seed}
}
\value{
Nothing (dumps statistics to screen)
}
\description{
Tests observed data (\code{ydat}) against a series of neutral models
}
\seealso{
\code{test1d}, \code{test2d}
}

