% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{run_tests}
\usage{
run_tests(size = 10, alpha = c(0.1, 0.1), nt = 100, ntests = 100,
  neutral = FALSE, ac_type = "moran", ymat, seed)
}
\arguments{
\item{size}{Size of the square grid on which to generate model. Total number
of points is size ^ 2. (Has no effect if \code{ymat} is given, in which case
value is determined by its dimensions)}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{nt}{Number of successive layers of temporal and spatial autocorrelation
used to generate test field}

\item{ntests}{Number of tests to run, with statistics calculated from the
mean of \code{ntests}}

\item{neutral}{If TRUE, uses a neutral field for the tests rather than a
structured field generated by \code{ives2d}}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{ymat}{Square matrix of observed values to be tested. If not given, will
be generated with specified values of \code{size} and \code{alpha}}

\item{seed}{Random seed}
}
\value{
Nothing (dumps statistics to screen)
}
\description{
Tests observed data (\code{ymat}) against a series of neutral models
}
\examples{
\dontrun{
run_tests () # should give low p-values
run_tests (neutral=TRUE) # should give high p-values
}

}
\seealso{
\code{test1d}, \code{test2d}
}

