% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test1d.R
\name{test1d}
\alias{test1d}
\title{test1d}
\usage{
test1d(ymat, alpha = c(0.1, 0.1), ntests = 100, actype = "moran",
  plot = FALSE)
}
\arguments{
\item{ymat}{Square matrix of observed values to be tested}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{ntests}{Number of repeats of neutral model used to calculate mean
rank--scale distribution}

\item{actype}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{plot}{If TRUE, produces a plot of rank--scale distributions}
}
\value{
A vector of four values as estimated by the neutral model:
\enumerate{
  \item Temporal autocorrelation coefficient
  \item Spatial autocorrelation coefficient
  \item Number of successive layers of spatio-temporal autocorrelation
  required to reproduce statistical properties of observed data
  \item Absolute difference between observed and modelled rank--scale
  distributions
}
}
\description{
Tests an observed matrix of data (\code{ymat}) against a one-dimensional
neutral model
}
\examples{
\dontrun{
alpha <- c (0.1, 0.1)
ymat <- ives2D (size=10, nt=1000, sd0=0.1, alpha=alpha)
result <- test1d (ymat, alpha, ntests=10)
}

}
\seealso{
\code{test2d}
}

