% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2d.R
\name{test2d}
\alias{test2d}
\title{test2d}
\usage{
test2d(z, nbs, alpha = c(0.1, 0.1), ntests = 100, ac_type = "moran",
  verbose = FALSE, plot = FALSE)
}
\arguments{
\item{z}{Vector of observed values to be tested}

\item{nbs}{An \code{spdep} \code{nb} object listing all neighbours of each
point}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{ntests}{Number of repeats of neutral model used to calculate mean
rank--scale distribution}

\item{ac_type}{type of autocorrelation statistic to use in tests
(\code{moran}, \code{geary}, or \code{getis-ord}=\code{go})}

\item{verbose}{If TRUE, dump progress details to screen}

\item{plot}{If TRUE, produces a plot of rank--scale distributions}
}
\value{
A vector of four values as estimated by the neutral model:
\enumerate{
  \item Temporal autocorrelation coefficient
  \item Spatial autocorrelation coefficient
  \item Number of successive layers of spatio-temporal autocorrelation
  required to reproduce statistical properties of observed data
  \item Absolute difference between observed and modelled rank--scale
  distributions
}
}
\description{
Tests an observed matrix of data (\code{ymat}) against a two-dimensional
neutral model
}
\examples{
\dontrun{
alpha <- c (0.1, 0.1)
ymat <- ives2d (size=10, nt=1000, sd0=0.1, alpha=alpha)
result <- test2d (ymat, alpha=alpha, ntests=10)
}

}
\seealso{
\code{test1d}
}

