% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2d.R
\name{test2d}
\alias{test2d}
\title{test2d}
\usage{
test2d(ydat, alpha = c(0.1, 0.1), ntests = 100, plot = FALSE)
}
\arguments{
\item{ydat}{Square matrix of observed values to be tested}

\item{alpha}{Vector of two components providing starting values for the
strength of autocorrelation in time and space}

\item{ntests}{Number of repeats of neutral model used to calculate mean
rank--scale distribution}

\item{plot}{If TRUE, produces a plot of rank--scale distributions}
}
\value{
A vector of four values as estimated by the neutral model:
\enumerate{
  \item Temporal autocorrelation coefficient
  \item Spatial autocorrelation coefficient
  \item Number of successive layers of spatio-temporal autocorrelation
  required to reproduce statistical properties of observed data
  \item Absolute difference between observed and modelled rank--scale
  distributions
}
}
\description{
Tests an observed matrix of data (\code{ydat}) against a two-dimensional
neutral model
}
\examples{
\dontrun{
alpha <- c (0.1, 0.1)
ydat <- ives2D (size=10, nt=1000, sd0=0.1, alpha=alpha)
result <- test2d (ydat, alpha=alpha, ntests=10)
}

}
\seealso{
\code{test1d}
}

